support/targets: Use target name as alias for run
We have a combination of supporting target-run and run-target, but actually the
most common use case for specifying a target name is to run that target.

Alias this to the run command, to support:
  make salvator-xs

which will halt the target, build the kernel and dtb, and set up the tftp
before re-powering the target.

Signed-off-by: Kieran Bingham <kieran@bingham.xyz>